Extend `ManagementAccess` api with SSH server configuration
This change extends the `ManagementAccess` to allow configuration of the
ssh server. It additionally removes the previously unused
`MinSampleInterval` configuration for the gnmi server.

+++ internal/provider/cisco/nxos/term.go
@@ -48,7 +48,7 @@ type VTY struct {
 	} `json:"execTmeout-items"`
 	SsLmtItems struct {
 		// Maximum number of concurrent vsh sessions.
-		SesLmt int `json:"sesLmt"`
+		SesLmt int8 `json:"sesLmt"`
 	} `json:"ssLmt-items"`
 }
